1. Refactored and moves rollsite2 to rollsite1.
