+++ jvm/roll_site/main/scala/com/webank/eggroll/rollsite/EggSiteBootstrap.scala
@@ -7,6 +7,6 @@ class EggSiteBootstrap extends BootstrapBase {
   override def init(args: Array[String]): Unit = this.args = args
 
   override def start(): Unit = {
-    Proxy.main(args)
+   // Proxy.main(args)
   }
 }
